+++ koupleless-common/src/main/java/com/alipay/sofa/koupleless/common/api/KouplelessScheduledExecutorService.java
@@ -14,82 +14,104 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.koupleless.common.util;
-
-import com.alipay.sofa.koupleless.common.api.KouplelessCallable;
-import com.alipay.sofa.koupleless.common.api.KouplelessRunnable;
+package com.alipay.sofa.koupleless.common.api;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 
 /**
  * @author lianglipeng.llp@alibaba-inc.com
- * @version $Id: KouplelessExecutorServiceAdaptor.java, v 0.1 2024年05月13日 12:03 立蓬 Exp $
+ * @version $Id: KouplelessScheduledExecutorServiceAdaptor.java, v 0.1 2024年05月13日 15:54 立蓬 Exp $
  */
-public class KouplelessExecutorService implements ExecutorService {
-    private final ExecutorService executorService;
+public class KouplelessScheduledExecutorService implements ScheduledExecutorService {
+    private final ScheduledExecutorService scheduledExecutorService;
 
-    public KouplelessExecutorService(ExecutorService executorService) {
-        this.executorService = executorService;
+    public KouplelessScheduledExecutorService(ScheduledExecutorService scheduledExecutorService) {
+        this.scheduledExecutorService = scheduledExecutorService;
+    }
+
+    @Override
+    public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit) {
+        return scheduledExecutorService.schedule(KouplelessRunnable.wrap(command), delay, unit);
+    }
+
+    @Override
+    public <V> ScheduledFuture<V> schedule(Callable<V> callable, long delay, TimeUnit unit) {
+        return scheduledExecutorService.schedule(KouplelessCallable.wrap(callable), delay, unit);
+    }
+
+    @Override
+    public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period,
+                                                  TimeUnit unit) {
+        return scheduledExecutorService.scheduleAtFixedRate(KouplelessRunnable.wrap(command),
+            initialDelay, period, unit);
+    }
+
+    @Override
+    public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay,
+                                                     long delay, TimeUnit unit) {
+        return scheduledExecutorService.scheduleWithFixedDelay(KouplelessRunnable.wrap(command),
+            initialDelay, delay, unit);
     }
 
     @Override
     public void shutdown() {
-        executorService.shutdown();
+        scheduledExecutorService.shutdown();
     }
 
     @Override
     public List<Runnable> shutdownNow() {
-        return executorService.shutdownNow();
+        return scheduledExecutorService.shutdownNow();
     }
 
     @Override
     public boolean isShutdown() {
-        return executorService.isShutdown();
+        return scheduledExecutorService.isShutdown();
     }
 
     @Override
     public boolean isTerminated() {
-        return executorService.isTerminated();
+        return scheduledExecutorService.isTerminated();
     }
 
     @Override
     public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
-        return executorService.awaitTermination(timeout, unit);
+        return scheduledExecutorService.awaitTermination(timeout, unit);
     }
 
     @Override
     public <T> Future<T> submit(Callable<T> task) {
-        return executorService.submit(KouplelessCallable.wrap(task));
+        return scheduledExecutorService.submit(KouplelessCallable.wrap(task));
     }
 
     @Override
     public <T> Future<T> submit(Runnable task, T result) {
-        return executorService.submit(KouplelessRunnable.wrap(task), result);
+        return scheduledExecutorService.submit(KouplelessRunnable.wrap(task), result);
     }
 
     @Override
     public Future<?> submit(Runnable task) {
-        return executorService.submit(KouplelessRunnable.wrap(task));
+        return scheduledExecutorService.submit(KouplelessRunnable.wrap(task));
     }
 
     @Override
     public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks) throws InterruptedException {
-        return executorService
+        return scheduledExecutorService
             .invokeAll(tasks.stream().map(KouplelessCallable::wrap).collect(Collectors.toList()));
     }
 
     @Override
     public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks, long timeout,
                                          TimeUnit unit) throws InterruptedException {
-        return executorService.invokeAll(
+        return scheduledExecutorService.invokeAll(
             tasks.stream().map(KouplelessCallable::wrap).collect(Collectors.toList()), timeout,
             unit);
     }
@@ -97,7 +119,7 @@ public class KouplelessExecutorService implements ExecutorService {
     @Override
     public <T> T invokeAny(Collection<? extends Callable<T>> tasks) throws InterruptedException,
                                                                     ExecutionException {
-        return executorService
+        return scheduledExecutorService
             .invokeAny(tasks.stream().map(KouplelessCallable::wrap).collect(Collectors.toList()));
     }
 
@@ -105,13 +127,13 @@ public class KouplelessExecutorService implements ExecutorService {
     public <T> T invokeAny(Collection<? extends Callable<T>> tasks, long timeout,
                            TimeUnit unit) throws InterruptedException, ExecutionException,
                                           TimeoutException {
-        return executorService.invokeAny(
+        return scheduledExecutorService.invokeAny(
             tasks.stream().map(KouplelessCallable::wrap).collect(Collectors.toList()), timeout,
             unit);
     }
 
     @Override
     public void execute(Runnable command) {
-        executorService.execute(KouplelessRunnable.wrap(command));
+        scheduledExecutorService.execute(KouplelessRunnable.wrap(command));
     }
 }
